// frontend-angular/src/app/core/generated-api/model/generoPaciente.ts
/**
 * API - Sistema de Citas Medicas
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Genero del paciente
 */
export const GeneroPaciente = {
    Masculino: 'Masculino',
    Femenino: 'Femenino',
    Otro: 'Otro',
} as const;
export type GeneroPaciente = typeof GeneroPaciente[keyof typeof GeneroPaciente];