// frontend-angular/src/app/core/generated-api/model/paciente.ts
/**
 * API - Sistema de Citas Medicas
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GeneroPaciente } from './generoPaciente';


export interface Paciente { 
    readonly id: number;
    nombreCompleto: string;
    documentoIdentidad: string;
    fechaNacimiento: string;
    telefono?: string;
    email?: string;
    direccion?: string;
    activo?: boolean;
    genero?: GeneroPaciente;
    /**
     * Texto libre para registrar alergias u otras condiciones de alerta medica del paciente
     */
    alergias?: string | null;
}
export namespace Paciente {
}